working update

+++ src/attacus/flutter/components/mouse_input.h
@@ -15,13 +15,13 @@ public:
     MouseInput(FlutterView& view);
     void Create();
     bool Dispatch(SDL_Event &e) override;
-    bool UpdatePointer(FlutterPointerPhase phase, double x, double y, size_t timestamp);
+    bool UpdatePointer(FlutterPointerPhase phase, float x, float y, size_t timestamp);
     // Accessors
     // Data members
     bool entered_ = false;
     bool mouseDown = false;
     int mouseId = 0;
-    int lastMouseX = 0, lastMouseY = 0;
+    float lastMouseX = 0.0f, lastMouseY = 0.0f;
     int64_t buttons_ = 0;
 };
